Changed ACPI architecture to namespaces
Converted ACPI structure to a set of namespaces for each table that contain simple "get" functions that read the appropriate info from offsets based on the table's base address. Information is no longer stored in structs, and are replace with these read functions. The struct implementation will be transitioned to the extended APIS repository.