Add intrinsics for EAPIs phys_mtrr class
- Add an overload to x64::is_physical_address_valid that takes
  the size of a physical address as an argument, as opposed to
  reading straight from cpuid.

- Define the mtrr::range structure that provides an API for
  constructing and querying MTRR ranges. It also ensures that
  the specific alignment and size requirements are met so that only
  valid MTRR ranges are created.

The phys_mtrr class in the the extended APIs need to use
bsf and popcnt to quickly determine if the variable range
MTRRs overlap a given address.

- Add _bsf and _popcnt intrinsics
- Add unit test stubs for _bsf and _popcnt

There are a variable number of variable ranges supported by
a given system, as reported by the IA32_MTRRCAP msr. Because
of the dynamic nature of the ranges, it is not useful to have
a separate physbase{n}, phymask{n} namespaces for each n that
is supported.

The fields within each variable MTRR has the same format, so it
is much cleaner and simpler to have one physbase and physmask
namespace for all of them.

- Remove the physbase[0-9], physmask[0-9] namespaces.
- Add single physbase and physmask namespaces for parsing
  all variable range MTRRs

Signed-off-by: Connor Davis <davisc@ainfosec.com>