interrupt_window: add post_external_interrupt
The queue_external_interrupt function enables external-interrupt exiting
which requires a vmwrite. For this to work properly, the vmcs the
vmwrite references must be the current vmcs. So if an interrupt needs to
be queued from a different vcpu, the caller must load() the target vcpu
first when using queue_external_itnerrupt.

- Add post_external_interrupt to queue the interrupt without
  accessing the vmcs. This enables "foreign" vcpu's to queue
  interrupts without loading the targeted one first.

- Give interrupt_window a std::mutex to synchronize accesses to its
  interrupt_queue.

Signed-off-by: Connor Davis <davisc@ainfosec.com>