Add BF_REQUEST_INIT_XUE entry point
BF_REQUEST_INIT_XUE instructs the VMM to "copy" the struct xue given
as an argument. The argument originates and is owned by the kernel
driver, which enables the device at driver load time. Once the driver is
loaded, the cable can be plugged in. Assuming the host successfully
configures the device, it may be read from using /dev/ttyUSB*.

BF_REQUEST_INIT_XUE gives the VMM a chance to map the physical pages
associated with the device into its own address space. Because the
device is already initialized at this point, the VMM need only provide
the xue_ops required by xue_write, which atm is memcpy and
virt_to_phys.

Signed-off-by: Connor Davis <davisc@ainfosec.com>