Use add_suproject for bfdriver
The makefiles in the bfdriver need to be configured by cmake
before building can happen. The two config items currently used
are BFVMM_XSAVE, which tells what level of AVX the VMM is compiled
for, and XUE_DIR, which is the directory to the xue debugger

add_subproject is needed so that the Makefile is configured prior
to the driver_* targets running. This enables downstream drivers
like microv to specify bfdriver as a dependency as well.