Fix event word link bug
Linux processes events on a fifo event channel by first looking
at the set of queues that need servicing. This set is equal to the
the set bits in evtchn_fifo_control_block->ready (each set bit
represents a queue at a given priority). For a queue that is ready
with priority p, linux starts consuming events referenced by the
control_block->head[p] port, and continues consuming each linked
event from the head. Here 'consuming' means that the link field is
cleared, and if the pending bit was set and mask bit clear, the
IRQ that the event maps to is handled.

Because linux clears each link, microv must ensure that the word
is reachable from control_block->head[p] whenever an upcall is
performed on the corresponding event channel.