Add acquire/release semantics for guest vcpus
It is possible for an interrupt bound to a guest vcpu to arrive
on a core that is different than the one the guest itself is running
on. A good example is when the host-resident microv driver receives
an interrupt on behalf of the guest. This driver often will not
be running on the same core as the guest because the IRQ affinity
logic in the kernel will route the interrupt to the core(s)
of its choosing.

In this situation, the VMM still needs to queue the interrupt into the
guest. The way this is done now is with guest->push_external_interrupt,
where guest is a pointer to the vcpu. However, this access races with the
destruction of the vcpu, which can happen at any time.

To address this, two functions are added to the bfmanager class:
    - acquire(id): get a pointer to the thing referenced by id. The
                   pointer returned is guaranteed to refer to a valid
                   object until the subsequent release(id) call.
    - release(id): release the reference previously acquired through
                   acquire.