Add invept to bfvmm::intel_x64::vcpu
Provide an API to invalidate a vcpu's EPT. This function passes the
value of the vcpu's vmcs EPTP field as the invalidation descriptor to a
single-context invalidation. However the vmcs itself is not accessed.
The actual EPTP value used is the same as the value written on the
previous ept_handler::set_eptp (which should equal what's in the vmcs
assuming that API is used). This means the vcpu's vmcs doesn't need to be
load()ed prior to invept(), but invept() should still be run on the physical
cpu(s) the vcpu is running on.