deps/hypervisor: Add EPT snoop bit
When EPT is shared with the IOMMU, bit 11 in each entry that maps
a page is interpreted by the IOMMU as a "snoop control" bit. This bit
tells the IOMMU whether the access to the page referenced by the entry
should snoop the CPU caches. This is useful in ensuring that DMA data is
consistent with main memory when the EPT memory type is write-back.

Note that the CPU's MMU ignores bit 11, so setting this bit has no
affect on the CPU side.