vmm: Add read/write CR8 handlers in guest vcpus
Add handlers for guest reads and writes to CR8. Prior to this change,
the guest could have written directly to CR8 in hardware. This patch
prevents that by enabling CR8 load and store exiting in each guest vcpu.

In addition, the handlers return false if an "unsupported" value is
written to CR8. If we ever run into a guest that uses CR8 for real,
we will need to incorporate the written values into the interrupt
injection logic. Right now values of 0 or 1 are allowed, and we simply
store this into a shadow variable in the vcpu. This requires no change
in the current interrupt injection code, because both PCI passthrough
and Xen callback vectors belong in a task priority class greater than 1.