loader/x64: Split loader_gdt.h into amd/intel directories
The Intel VMCS fields related to the GDT differs slightly from the AMD
VMCB in that the base of the TSS must be initialized properly in the
VMCS_GUEST_TR_BASE field, so get_segment_descriptor_base needs to be
system-descriptor aware so that it can return a 64-bit base address
for the TSS.

Also each segment descriptor's access rights need to include the 4-bit,
portion of the segment limit in between bits 11:8 and 7:0 of the actual
attributes (the AMD code scrunches the twelve access-rights bits into a
12-bit field, thereby omitting the 4 limit bits).

The loader_gdt.h under x64/intel from this patch addresses these two
concerns.  The old loader_gdt.h has been moved to x64/amd with no
funcitional change.